requirements.txt
faiss-cpu==1.8.0
fastapi==0.110.2
gradio==4.27.0
httpx==0.27.0
numpy==1.26.4
python-dotenv==1.0.1
uvicorn==0.29.0